Enforce specific version of Habitat
Ensure that the installed version of the hab toolchain is >= to the
required version.

Given the rapid development and release cadence of Habitat, we want to
ensure that developers are using a version of Habitat is is well tested
against both this library and the supporting automation (i.e. Expeditor)

Signed-off-by: Tom Duffield <tom@chef.io>