Fix export_docker_image to fail if error found
Before we were not verifying that when we build a hab package, if the
build process failed we were continuing with the pkg export, now we are
erroring out.

Also if we fail tryingt to export, lets print out the command we tried
to run so that the user can diagnose because definitely something went
wrong.

Signed-off-by: Salim Afiune <afiune@chef.io>