Use 'pwd' instead of Travis build directory to determine location
The go language in Travis actually symlinks code into the GOPATH. By
cd-ing out of the GOPATH, we could be causing unknown behavior. This
change just determines where we are starting out. This has the added
benefit of expanding support outside of Travis by no longer depending
on Travis variables.

Signed-off-by: Tom Duffield <tom@chef.io>