dep_ensure runs only when vendor/ is missing
This commit is updating the `dep_ensure` helper to only run when the
vendor/ directory is missing. Think about this helper as a safe hook
that, when you call it, it will ensure you have your dependencies, but
if you already have them, then it wonth modify them. If you need to
actually update your Go dependencies you have to run `dep ensure`
manually.

Signed-off-by: Salim Afiune <afiune@chef.io>