install_if_missing, just do it quietly
There are many places in the A2 .studiorc and friends that we use
`install_if_missing` deep inside of functions and to wrap other
tools. We rely on `install_if_missing` to work silently, because
we rely on having the ability to parse the output of the tools that
it installs. We need to make `install_if_missing` quiet again.

This change landed in stable today and has caused multiple things
to break:

https://github.com/chef/ci-studio-common/commit/4aab13426154cb2afff21536d0973685caa83cfd

We've already shipped one fix to A2, but it looks like this is the
root cause so it's better to just make this silent like it's always
been than to play whack-a-mole for every place we use it.

Signed-off-by: Stephen Delano <stephen@chef.io>