Update generate_netrc_config error message
With the new version of habitat, we no longer use the `.secrets` pattern but instead we use the proper habitat way to share secrets through env variables, in our case we use `HAB_STUDIO_SECRET_GITHUB_TOKEN` that needs to be exported before entering the studio. 

This PR is updating the error message to give better instructions of what needs to happen to fix the problem.