2.0 Release! Rewrite in Go
To better handle support for multiple platforms, we're starting to
rewrite a number of our utilities in Go. This will make it easier to
support for multiple platforms (linux, darwin, and windows).

All bash utilities have been removed from this repo. A select few
utilities focusing on omnibus have been moved to chef/omnibus-toolchain.

We've replaced a number of utilities with Go counterparts:
  * ci-studio-common-util
  * did-modify
  * file-mod (no more perl requirement!)
  * install-buildkite-agent (no longer runs curl|bash)
  * install-habitat
  * purge-habitat (moved into install-habitat)

We've removed the following (unused) buildkite agent hooks:
  * remove-containers (unused)
  * use-merge-head (this is a buggy feature from GitHub)
  * workspace-reassing-ownership (this is now handled upstream)
  * expeditor (now called ci-studio-common)

We added a new utility called vault-util that can load secrets from
Vault in a number of common configurations to make account management
easier on your CI instances.

Signed-off-by: Tom Duffield <tom@chef.io>