Cache GitHub token on disk to avoid spamming token creation
The way this utility was previously designed, it would create a fresh
installation token on every git request, which could be more than 1 per
second. That would essentially create one token for every git request,
which is not what we want.

This design will create a token and save it to disk. We'll then read
the token from disk if it exists and is unexpired, otherwise we'll
create a fresh token.

We are writing the token to disk in plaintext, but it is a temporary
token. Given the current usage patterns, it's an acceptable trade off
at the moment. In the future, if we become more concerned, we can store
the token encrypted. However, we store AWS credentials in plaintext on
disk, so I think we'll be okay.

Signed-off-by: Tom Duffield <tom@chef.io>